Begin transition to lexer goroutine.
This breaks nex programs that have nested regexes containing return statements
in their actions.

I simplified the generated code. Instead of maintaining a stack of states for
each level of nesting, we use Go functions instead, and let the language take
care of it.

However, return statements in the actions will destroy this state. My original
solution was to maintain my own stack that was preserved between calls to the
lexer. I now intend to solve this problem by running the lexer in its own
goroutine.