rewrite binary finder; make windows friendly
Specifically, the code assumed forward slash path separators which fails
on windows. Changed to use exec.LookPath which does a better job of
searching for binaries, and added support for nex to be anywhere in PATH.

Since the cp command does not exist on windows, this also changes one
command to copy the files directly instead of execing cp.